Correct the middleware to use array ref. But this does let me rethink about this header thing.
If it beccomes an array ref, it's hard for middleware to add or update
headers. They need to iterate the array ref to see if it's already
there, or otherwise push.

Rack's joining with "\n" does feel dirty, but that solves this exact
problem. Another option would be to require/allow array ref as a
value, but that's a little confusing with multiple values that should
be joined with ", " (in the application).